+++ CHANGELOG.md
@@ -32,7 +32,7 @@
 
 Initial release
 
-[unreleased]: https://github.com/blackbox-log/blackbox-log/compare/v0.1.3...HEAD
+[unreleased]: https://github.com/blackbox-log/blackbox-log-ts/compare/v0.1.3...HEAD
 [0.1.3]: https://github.com/blackbox-log/blackbox-log-ts/compare/v0.1.2...v0.1.3
 [0.1.2]: https://github.com/blackbox-log/blackbox-log-ts/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/blackbox-log/blackbox-log-ts/compare/v0.1.0...v0.1.1