+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [0.1.4] - 2023-05-10
+
 ### Added
 
 -   [`Stats.progress`](https://docs.rs/blackbox-log/0.3.1/blackbox_log/data/struct.Stats.html#structfield.progress)
@@ -32,7 +34,8 @@
 
 Initial release
 
-[unreleased]: https://github.com/blackbox-log/blackbox-log-ts/compare/v0.1.3...HEAD
+[unreleased]: https://github.com/blackbox-log/blackbox-log-ts/compare/v0.1.4...HEAD
+[0.1.4]: https://github.com/blackbox-log/blackbox-log-ts/compare/v0.1.3...v0.1.4
 [0.1.3]: https://github.com/blackbox-log/blackbox-log-ts/compare/v0.1.2...v0.1.3
 [0.1.2]: https://github.com/blackbox-log/blackbox-log-ts/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/blackbox-log/blackbox-log-ts/compare/v0.1.0...v0.1.1